refactor typo
